LPS-106031 Rename AssetService to InfoItemService
